requirements.txt
numpy==1.24.2
torchvision==0.14.1
tensorboard==2.12.0
matplotlib==3.7.1
opencv_python==4.7.0.72
pandas==1.5.3